+++ recipe_of_five_front/src/App.js
@@ -2,13 +2,13 @@
 import React, { createContext } from "react";
 import "bootstrap/dist/css/bootstrap.min.css";
 
-import SearchInd from "./components/SearchInd";
+import SearchInd from "./components/SearchInd/SearchInd";
 import Ranking from "./components/Ranking/Ranking";
 import RecipeDescrip from "./components/RecipeDesc/RecipeDescrip"
-import FilterBtn from "./components/FilterBtn";
+import FilterBtn from "./components/SearchInd/FilterBtn";
 import SearchProvider from "./context/SearchProvider";
-import Ingredients from "./components/Ingredients";
-import SendSearch from "./components/SendSearch";
+import Ingredients from "./components/SearchInd/Ingredients";
+import SendSearch from "./components/SearchInd/SendSearch";
 import RecipeResult from "./components/RecipeResult/RecipeResult";
