+++ recipe_of_five_front/src/components/SearchInd/FilterBtn.jsx
@@ -1,7 +1,7 @@
 import React, { useState, useContext } from "react";
 import ToggleButtonGroup from "react-bootstrap/ToggleButtonGroup";
 import ToggleButton from "react-bootstrap/ToggleButton";
-import { RecipeFilterContext } from "../context/SearchProvider";
+import { RecipeFilterContext } from "../../context/SearchProvider";
 
 export default function FilterBtn() {
   const { filterMethod, setFilterMethod } = useContext(RecipeFilterContext);
