+++ recipe_of_five_front/src/components/SearchInd/SendSearch.jsx
@@ -1,21 +1,18 @@
 import React, { useContext, useEffect } from "react";
 import { Button } from "react-bootstrap";
-import { RecipeFilterContext } from "../context/SearchProvider";
-import { Navigate, useNavigate } from "react-router-dom";
+import { RecipeFilterContext } from "../../context/SearchProvider";
 import axios from "axios";
 
 export default function SendSearch() {
-    const navigate = useNavigate();
-    const {
-        searchOptions,
-        setSearchOptions,
-        filterMethod,
-        include,
-        exclude,
-        ranking,
-        setRanking,
-    } = useContext(RecipeFilterContext);
-
+  const {
+    searchOptions,
+    setSearchOptions,
+    filterMethod,
+    include,
+    exclude,
+    ranking,
+    setRanking,
+  } = useContext(RecipeFilterContext);
 
   const handleSearch = () => {
     setSearchOptions({
@@ -27,7 +24,6 @@ export default function SendSearch() {
     //버튼 클릭시 세부 홈페이지로 이동 예정
     //navigate("/detail");
   };
-
 
   // 최초 랭킹 화면 get 요청
   useEffect(() => {
